Fixed test_rates_close()
This test used to check if the rates were less than $10 different on multiple
exchanges. As the price of Bitcoin Cash increased it became possible to have
different exhanges rate the price of a coin more than $10 above or below each
other.

test_rates_close() now checks if the rates are less than 10% different.