Replace IGNORED_ERRORS with appropriate Exceptions
We do not want to catch every single Excpetion, but all network errors
indicating that the server failed to provide a correct HTTP response.

I replaced the content of IGNORED_ERRORS with Exceptions found at:
https://2.python-requests.org/en/master/_modules/requests/exceptions/

I did not include the excpetions from the client (MissingSchema,
InvalidSchema, InvalidURL, InvalidHeader and so on...).